+++ HISTORY.md
@@ -1,6 +1,10 @@
 # 0.5.4
 
-Added a method, `DimensionalData.DimArray(::FlexiChain)`, to convert a `FlexiChain` into a 3D `DimArray` with dimensions (iters, chains, parameters).
+Added a method, `DimensionalData.DimArray(::FlexiChain, kwargs...)`, to convert a `FlexiChain` into a 3D `DimArray` with dimensions `(iters, chains, parameters)`.
+A method for `Base.Array` has also been added, which is the same but drops the dimensional metadata.
+
+Implemented `PosteriorStats.loo` for chains of log-likelihoods, as well as Turing.jl models.
+Please see the documentation for more information about the PosteriorStats integration.
 
 # 0.5.3